pyproject.toml
[project]
name = "fastapi-voyager"
dynamic = ["version"]
description = "Visualize FastAPI application's routing tree and dependencies"
authors = [ { name = "Tangkikodo", email = "allmonday@126.com" } ]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.10"
keywords = ["fastapi", "visualization", "routing", "openapi"]
dependencies = [
  "pydantic-resolve>=2.4.3",
  "jinja2>=3.0.0"
]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Framework :: FastAPI",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License"
]

[project.scripts]
voyager = "fastapi_voyager.cli:main"

[project.urls]
Homepage = "https://github.com/allmonday/fastapi-voyager"
Source = "https://github.com/allmonday/fastapi-voyager"

[project.optional-dependencies]
dev = ["uvicorn", "ruff", "pytest", "pytest-asyncio", "httpx", "fastapi>=0.110", "django-ninja", "litestar"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "src/fastapi_voyager/version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/fastapi_voyager"]

[tool.uv]
# You can pin resolution or indexes here later.

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[dependency-groups]
dev = [
    "django-ninja>=1.5.3",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "litestar>=2.19.0",
    "pytest-asyncio>=1.3.0",
]
